move progress guard to helpers, update lib
