+++ music_assistant/providers/audiobookshelf/helpers.py
@@ -1,7 +1,9 @@
 """Helpers for Audiobookshelf provider."""
 
+import time
 from dataclasses import dataclass, field
 
+from aioaudiobookshelf.schema.media_progress import MediaProgress
 from mashumaro.mixins.dict import DataClassDictMixin
 
 
@@ -22,3 +24,71 @@ class LibrariesHelper(DataClassDictMixin):
 
     audiobooks: dict[str, LibraryHelper] = field(default_factory=dict)
     podcasts: dict[str, LibraryHelper] = field(default_factory=dict)
+
+
+@dataclass(kw_only=True)
+class _ProgressHelper:
+    id_: str  # audiobook or podcast id
+    episode_id: str | None = None
+    last_update_ms: int  # last update in ms epoch (same as last_update in abs)
+
+
+class ProgressGuard:
+    """Class used to avoid ping pong."""
+
+    def __init__(self) -> None:
+        """Init."""
+        self._progresses: list[_ProgressHelper] = []
+        self._max_progresses = 100
+        # 12s have to have passed before we accept an external progress update
+        # abs updates every 15 s
+        self._min_time_between_updates_ms = 12000
+
+    def _get_progress(self, item_id: str, episode_id: str | None = None) -> _ProgressHelper | None:
+        for x in self._progresses:
+            if x.id_ == item_id and x.episode_id == episode_id:
+                return x
+        return None
+
+    def _remove_oldest(self) -> None:
+        progresses = sorted(self._progresses, key=lambda x: x.last_update_ms)
+        if len(progresses) > 0:
+            self._progresses.remove(progresses[0])
+
+    def remove_progress(self, item_id: str, episode_id: str | None = None) -> None:
+        """Remove a progress."""
+        progress = self._get_progress(item_id=item_id, episode_id=episode_id)
+        if progress is not None:
+            self._progresses.remove(progress)
+
+    def add_progress(self, item_id: str, episode_id: str | None = None) -> None:
+        """Store episode progress."""
+        if len(self._progresses) > self._max_progresses:
+            self._remove_oldest()
+        self.remove_progress(item_id=item_id, episode_id=episode_id)
+        progress = _ProgressHelper(
+            id_=item_id, episode_id=episode_id, last_update_ms=int(time.time() * 1000)
+        )
+        self._progresses.append(progress)
+
+    def guard_ok_abs(self, abs_progress: MediaProgress) -> bool:
+        """Check if progress update is ok."""
+        item_id = abs_progress.library_item_id
+        episode_id = abs_progress.episode_id
+        stored_progress = self._get_progress(item_id=item_id, episode_id=episode_id)
+        if stored_progress is None:
+            return True
+        return bool(
+            abs_progress.last_update - stored_progress.last_update_ms
+            >= self._min_time_between_updates_ms
+        )
+
+    def guard_ok_mass(self, item_id: str, episode_id: str | None = None) -> bool:
+        """Check if progress update is ok."""
+        stored_progress = self._get_progress(item_id=item_id, episode_id=episode_id)
+        if stored_progress is None:
+            return True
+        return (
+            int(time.time() * 1000) - stored_progress.last_update_ms
+            >= self._min_time_between_updates_ms
+        )
