+++ music_assistant/providers/audiobookshelf/manifest.json
@@ -7,7 +7,7 @@
     "@fmunkes"
   ],
   "requirements": [
-    "aioaudiobookshelf==0.1.1"
+    "aioaudiobookshelf==0.1.2"
   ],
   "documentation": "https://music-assistant.io/music-providers/audiobookshelf",
   "multi_instance": true
