+++ requirements_all.txt
@@ -1,7 +1,7 @@
 # WARNING: this file is autogenerated!
 
 Brotli>=1.0.9
-aioaudiobookshelf==0.1.1
+aioaudiobookshelf==0.1.2
 aiodns>=3.0.0
 aiofiles==24.1.0
 aiohttp==3.11.11
